[web] Rename LoadingEnvironment to Loading
